+++ src/plugins/bold/bold.test.js
@@ -8,8 +8,13 @@ describe('Bold plugin', () => {
 	let jodit;
 	beforeEach(() => {
 		jodit = getJodit();
+		jodit.focus();
 		jodit.value = '<p>|<br></p>';
 		setCursorToChar(jodit);
+	});
+
+	afterEach(() => {
+		jodit.destruct();
 	});
 
 	describe('Click bold button', () => {
@@ -18,6 +23,36 @@ describe('Bold plugin', () => {
 			jodit.async.requestIdlePromise();
 			replaceCursorToChar(jodit);
 			expect(jodit.value).eq('<p><strong>|</strong></p>');
+		});
+	});
+
+	describe('Click subscript button', () => {
+		it('Should add empty SUB element', async () => {
+			clickButton('subscript', jodit);
+			jodit.async.requestIdlePromise();
+			replaceCursorToChar(jodit);
+			expect(jodit.value).eq('<p><sub>|</sub></p>');
+		});
+
+		describe('Inside sub tag', () => {
+			it('Should move cursor in the outside from the tag', async () => {
+				jodit.value = '<p><sup>test|</sup></p>';
+				setCursorToChar(jodit);
+				clickButton('superscript', jodit);
+				jodit.async.requestIdlePromise();
+				jodit.s.insertHTML('pop');
+				replaceCursorToChar(jodit);
+				expect(jodit.value).eq('<p><sup>test</sup>pop|</p>');
+			});
+		});
+	});
+
+	describe('Click superscript button', () => {
+		it('Should add empty SUP element', async () => {
+			clickButton('superscript', jodit);
+			jodit.async.requestIdlePromise();
+			replaceCursorToChar(jodit);
+			expect(jodit.value).eq('<p><sup>|</sup></p>');
 		});
 	});
 });